+++ tests/utils/test_search_runs_utils.py
@@ -13,7 +13,11 @@ from mlflow.entities import (
     RunTag,
 )
 from mlflow.exceptions import MlflowException
-from mlflow.utils.search_utils import SearchUtils
+from mlflow.utils.search_runs_utils import SearchRunsUtils
+
+
+def _get_regex_error_message(error_message):
+    return r".*{}.*".format(error_message)
 
 
 @pytest.mark.parametrize(
@@ -112,37 +116,7 @@ from mlflow.utils.search_utils import SearchUtils
     ],
 )
 def test_filter_for_run(filter_string, parsed_filter):
-    assert SearchUtils.parse_filter_for_run(filter_string) == parsed_filter
-
-
-@pytest.mark.parametrize(
-    "filter_string, parsed_filter",
-    [
-        ("", []),
-        (
-            "model.name = 'mod'",
-            [{"type": "attribute", "comparator": "=", "key": "name", "value": "mod"}],
-        ),
-        (
-            "name ILIKE '%final new%'",
-            [{"type": "attribute", "comparator": "ILIKE", "key": "name", "value": "%final new%"}],
-        ),
-        (
-            "registered_model.name LIKE '%new%'",
-            [{"type": "attribute", "comparator": "LIKE", "key": "name", "value": "%new%"}],
-        ),
-        (
-            "tag.`training algorithm` != 'xgboost'",
-            [{"type": "tag", "comparator": "!=", "key": "training algorithm", "value": "xgboost"}],
-        ),
-        (
-            '`model_tag`.owner LIKE "%cc"',
-            [{"type": "tag", "comparator": "LIKE", "key": "owner", "value": "%cc"}],
-        ),
-    ],
-)
-def test_filter_for_registered_model(filter_string, parsed_filter):
-    assert SearchUtils.parse_filter_for_registered_models(filter_string) == parsed_filter
+    assert SearchRunsUtils.parse_filter_for_run(filter_string) == parsed_filter
 
 
 @pytest.mark.parametrize(
@@ -158,20 +132,7 @@ def test_filter_for_registered_model(filter_string, parsed_filter):
     ],
 )
 def test_correct_quote_trimming_for_run(filter_string, parsed_filter):
-    assert SearchUtils.parse_filter_for_run(filter_string) == parsed_filter
-
-
-@pytest.mark.parametrize(
-    "filter_string, parsed_filter",
-    [
-        ("model_tag.m = 'LR'", [{"type": "tag", "comparator": "=", "key": "m", "value": "LR"}]),
-        ('model_tag.m = "LR"', [{"type": "tag", "comparator": "=", "key": "m", "value": "LR"}]),
-        ('model_tag.m = "LR"', [{"type": "tag", "comparator": "=", "key": "m", "value": "LR"}]),
-        ('tag.m = "L\'Hosp"', [{"type": "tag", "comparator": "=", "key": "m", "value": "L'Hosp"}]),
-    ],
-)
-def test_correct_quote_trimming_for_registered_model(filter_string, parsed_filter):
-    assert SearchUtils.parse_filter_for_registered_models(filter_string) == parsed_filter
+    assert SearchRunsUtils.parse_filter_for_run(filter_string) == parsed_filter
 
 
 @pytest.mark.parametrize(
@@ -184,13 +145,13 @@ def test_correct_quote_trimming_for_registered_model(filter_string, parsed_filte
         ("attri.x != 1", "Invalid entity type"),
         ("a.x != 1", "Invalid entity type"),
         ("model >= 'LR'", "Invalid attribute key"),
-        ("metrics.A > 0.1 OR params.B = 'LR'", "Invalid clause(s) in filter string"),
-        ("metrics.A > 0.1 NAND params.B = 'LR'", "Invalid clause(s) in filter string"),
-        ("metrics.A > 0.1 AND (params.B = 'LR')", "Invalid clause(s) in filter string"),
-        ("`metrics.A > 0.1", "Invalid clause(s) in filter string"),
-        ("param`.A > 0.1", "Invalid clause(s) in filter string"),
-        ("`dummy.A > 0.1", "Invalid clause(s) in filter string"),
-        ("dummy`.A > 0.1", "Invalid clause(s) in filter string"),
+        ("metrics.A > 0.1 OR params.B = 'LR'", "Invalid clause\\(s\\) in filter string"),
+        ("metrics.A > 0.1 NAND params.B = 'LR'", "Invalid clause\\(s\\) in filter string"),
+        ("metrics.A > 0.1 AND (params.B = 'LR')", "Invalid clause\\(s\\) in filter string"),
+        ("`metrics.A > 0.1", "Invalid clause\\(s\\) in filter string"),
+        ("param`.A > 0.1", "Invalid clause\\(s\\) in filter string"),
+        ("`dummy.A > 0.1", "Invalid clause\\(s\\) in filter string"),
+        ("dummy`.A > 0.1", "Invalid clause\\(s\\) in filter string"),
         ("attribute.start != 1", "Invalid attribute key"),
         ("attribute.start_time != 1", "Invalid attribute key"),
         ("attribute.end_time != 1", "Invalid attribute key"),
@@ -203,48 +164,12 @@ def test_correct_quote_trimming_for_registered_model(filter_string, parsed_filte
         ("attribute._status != 'RUNNING'", "Invalid attribute key"),
         ("name != 1", "Invalid attribute key"),
         ("experiment_id != 1", "Invalid attribute key"),
-        ("attribute.status = true", "Invalid clause(s) in filter string"),
+        ("attribute.status = true", "Invalid clause\\(s\\) in filter string"),
     ],
 )
 def test_error_filter_for_run(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_run(filter_string)
-    assert error_message in e.value.message
-
-
-@pytest.mark.parametrize(
-    "filter_string, error_message",
-    [
-        (
-            "tag.owner = 'zero qu'; name = 'some mod'",
-            "Search filter contained multiple expression",
-        ),
-        ("metric.acc >= 0.94", "Invalid entity type"),
-        ("model.registeredModelId = 0000", "Invalid attribute key"),
-        ("param.model = 'LR'", "Invalid entity type"),
-        ("attri.x != 1", "Invalid entity type"),
-        ("tag.owner = 'zero qu' OR name = 'some mod'", "Invalid clause(s) in filter string"),
-        ("tag.owner = 'zero qu' NAND name = 'some mod'", "Invalid clause(s) in filter string"),
-        ("tag.owner = 'zero qu' AND (name = 'some mod')", "Invalid clause(s) in filter string"),
-        ("`tag.A = 'B'", "Invalid clause(s) in filter string"),
-        ("tag`.A = 'B'", "Invalid clause(s) in filter string"),
-        ("attribute.ID != '1'", "Invalid attribute key"),
-        ("attribute.stage != '1'", "Invalid attribute key"),
-        ("attribute.end_time != '1'", "Invalid attribute key"),
-        ("attribute.run_id != '1'", "Invalid attribute key"),
-        ("attribute.run_uuid != '1'", "Invalid attribute key"),
-        ("attribute._status != 'RUNNING'", "Invalid attribute key"),
-        ("attribute.run_id != '1'", "Invalid attribute key"),
-        ("attribute.run_uuid != '1'", "Invalid attribute key"),
-        ("run_id != '1'", "Invalid attribute key"),
-        ("stage != '1'", "Invalid attribute key"),
-        ("attribute.status = true", "Invalid clause(s) in filter string"),
-    ],
-)
-def test_error_filter_for_registered_model(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_registered_models(filter_string)
-    assert error_message in e.value.message
+    with pytest.raises(MlflowException, match=_get_regex_error_message(error_message)):
+        SearchRunsUtils.parse_filter_for_run(filter_string)
 
 
 @pytest.mark.parametrize(
@@ -261,29 +186,8 @@ def test_error_filter_for_registered_model(filter_string, error_message):
     ],
 )
 def test_error_comparison_clauses_for_run(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_run(filter_string)
-    assert error_message in e.value.message
-
-
-@pytest.mark.parametrize(
-    "filter_string, error_message",
-    [
-        ("tag.acc = 5", "Expected a quoted string value for tag"),
-        ("tags.acc = 5", "Expected a quoted string value for tag"),
-        (
-            "tag.model != tag.model",
-            "Parameter value is either not quoted or unidentified quote types",
-        ),
-        ("1.0 > tag.acc", "Expected 'Identifier' found"),
-        ("attribute.name = 1", "Expected a quoted string value for attributes"),
-        ("name = 1", "Expected a quoted string value for attributes"),
-    ],
-)
-def test_error_comparison_clauses_for_registered_model(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_registered_models(filter_string)
-    assert error_message in e.value.message
+    with pytest.raises(MlflowException, match=_get_regex_error_message(error_message)):
+        SearchRunsUtils.parse_filter_for_run(filter_string)
 
 
 @pytest.mark.parametrize(
@@ -292,76 +196,36 @@ def test_error_comparison_clauses_for_registered_model(filter_string, error_mess
         ("params.acc = LR", "value is either not quoted or unidentified quote types"),
         ("tags.acc = LR", "value is either not quoted or unidentified quote types"),
         ("params.acc = `LR`", "value is either not quoted or unidentified quote types"),
-        ("params.'acc = LR", "Invalid clause(s) in filter string"),
-        ("params.acc = 'LR", "Invalid clause(s) in filter string"),
-        ("params.acc = LR'", "Invalid clause(s) in filter string"),
-        ("params.acc = \"LR'", "Invalid clause(s) in filter string"),
-        ("tags.acc = \"LR'", "Invalid clause(s) in filter string"),
-        ("tags.acc = = 'LR'", "Invalid clause(s) in filter string"),
-        ("attribute.status IS 'RUNNING'", "Invalid clause(s) in filter string"),
-        ("status IS 'RUNNING'", "Invalid clause(s) in filter string"),
+        ("params.'acc = LR", "Invalid clause\\(s\\) in filter string"),
+        ("params.acc = 'LR", "Invalid clause\\(s\\) in filter string"),
+        ("params.acc = LR'", "Invalid clause\\(s\\) in filter string"),
+        ("params.acc = \"LR'", "Invalid clause\\(s\\) in filter string"),
+        ("tags.acc = \"LR'", "Invalid clause\\(s\\) in filter string"),
+        ("tags.acc = = 'LR'", "Invalid clause\\(s\\) in filter string"),
+        ("attribute.status IS 'RUNNING'", "Invalid clause\\(s\\) in filter string"),
+        ("status IS 'RUNNING'", "Invalid clause\\(s\\) in filter string"),
     ],
 )
 def test_bad_quotes_for_run(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_run(filter_string)
-    assert error_message in e.value.message
+    with pytest.raises(MlflowException, match=_get_regex_error_message(error_message)):
+        SearchRunsUtils.parse_filter_for_run(filter_string)
 
 
 @pytest.mark.parametrize(
     "filter_string, error_message",
     [
-        ("tags.acc = LR", "value is either not quoted or unidentified quote types"),
-        ("tags.acc = `LR`", "value is either not quoted or unidentified quote types"),
-        ("tags.'acc = LR", "Invalid clause(s) in filter string"),
-        ("tags.acc = 'LR", "Invalid clause(s) in filter string"),
-        ("tags.acc = LR'", "Invalid clause(s) in filter string"),
-        ("tags.acc = \"LR'", "Invalid clause(s) in filter string"),
-        ("tags.acc = = 'LR'", "Invalid clause(s) in filter string"),
-        ("attribute.name IS 'some mod'", "Invalid clause(s) in filter string"),
-        ("name IS 'some mod'", "Invalid clause(s) in filter string"),
-    ],
-)
-def test_bad_quotes_for_registered_model(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_registered_models(filter_string)
-    assert error_message in e.value.message
-
-
-@pytest.mark.parametrize(
-    "filter_string, error_message",
-    [
-        ("params.acc LR !=", "Invalid clause(s) in filter string"),
-        ("params.acc LR", "Invalid clause(s) in filter string"),
-        ("metric.acc !=", "Invalid clause(s) in filter string"),
+        ("params.acc LR !=", "Invalid clause\\(s\\) in filter string"),
+        ("params.acc LR", "Invalid clause\\(s\\) in filter string"),
+        ("metric.acc !=", "Invalid clause\\(s\\) in filter string"),
         ("acc != 1.0", "Invalid attribute key"),
-        ("foo is null", "Invalid clause(s) in filter string"),
+        ("foo is null", "Invalid clause\\(s\\) in filter string"),
         ("1=1", "Expected 'Identifier' found"),
         ("1==2", "Expected 'Identifier' found"),
     ],
 )
 def test_invalid_clauses_for_run(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_run(filter_string)
-    assert error_message in e.value.message
-
-
-@pytest.mark.parametrize(
-    "filter_string, error_message",
-    [
-        ("tags.acc LR !=", "Invalid clause(s) in filter string"),
-        ("tags.acc LR", "Invalid clause(s) in filter string"),
-        ("tags.acc !=", "Invalid clause(s) in filter string"),
-        ("acc != 1.0", "Invalid attribute key"),
-        ("foo is null", "Invalid clause(s) in filter string"),
-        ("1=1", "Expected 'Identifier' found"),
-        ("1==2", "Expected 'Identifier' found"),
-    ],
-)
-def test_invalid_clauses_for_registered_model(filter_string, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_filter_for_registered_models(filter_string)
-    assert error_message in e.value.message
+    with pytest.raises(MlflowException, match=_get_regex_error_message(error_message)):
+        SearchRunsUtils.parse_filter_for_run(filter_string)
 
 
 @pytest.mark.parametrize(
@@ -397,9 +261,8 @@ def test_bad_comparators_for_run(entity_type, bad_comparators, key, entity_value
             bad_filter = "{entity_type}.{key} {comparator} {value}".format(
                 entity_type=entity_type, key=key, comparator=bad_comparator, value=entity_value,
             )
-        with pytest.raises(MlflowException) as e:
-            SearchUtils.filter_runs([run], bad_filter)
-        assert "Invalid comparator" in str(e.value.message)
+        with pytest.raises(MlflowException, match=_get_regex_error_message("Invalid comparator")):
+            SearchRunsUtils.filter_runs([run], bad_filter)
 
 
 @pytest.mark.parametrize(
@@ -470,7 +333,7 @@ def test_correct_filtering_for_runs(filter_string, matching_runs):
             ),
         ),
     ]
-    filtered_runs = SearchUtils.filter_runs(runs, filter_string)
+    filtered_runs = SearchRunsUtils.filter_runs(runs, filter_string)
     assert set(filtered_runs) == set([runs[i] for i in matching_runs])
 
 
@@ -545,7 +408,7 @@ def test_correct_sorting_for_runs(order_bys, matching_runs):
             ),
         ),
     ]
-    sorted_runs = SearchUtils.sort_runs(runs, order_bys)
+    sorted_runs = SearchRunsUtils.sort_runs(runs, order_bys)
     sorted_run_indices = []
     for run in sorted_runs:
         for i, r in enumerate(runs):
@@ -573,8 +436,8 @@ def test_order_by_search_run_metric_with_nans_and_infs():
         )
         for x in metric_vals_str
     ]
-    sorted_runs_asc = [x.info.run_id for x in SearchUtils.sort_runs(runs, ["metrics.x asc"])]
-    sorted_runs_desc = [x.info.run_id for x in SearchUtils.sort_runs(runs, ["metrics.x desc"])]
+    sorted_runs_asc = [x.info.run_id for x in SearchRunsUtils.sort_runs(runs, ["metrics.x asc"])]
+    sorted_runs_desc = [x.info.run_id for x in SearchRunsUtils.sort_runs(runs, ["metrics.x desc"])]
     # asc
     assert ["-inf", "-1000", "0", "1000", "inf", "nan"] == sorted_runs_asc
     # desc
@@ -604,41 +467,8 @@ def test_order_by_search_run_metric_with_nans_and_infs():
     ],
 )
 def test_invalid_order_by_search_runs(order_by, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_order_by_for_search_runs(order_by)
-    assert error_message in e.value.message
-
-
-@pytest.mark.parametrize(
-    "order_by, error_message",
-    [
-        ("m.acc", "Invalid entity type"),
-        ("acc", "Invalid attribute key"),
-        ("attri.x", "Invalid entity type"),
-        ("`tags.A", "Invalid order_by clause"),
-        ("`tags.A`", "Invalid entity type"),
-        ("attribute.start", "Invalid attribute key"),
-        ("attribute.stage", "Invalid attribute key"),
-        ("start", "Invalid order_by clause"),
-        ("tags.A != 1", "Invalid order_by clause"),
-        ("attribute.name ACS", "Invalid ordering key"),
-        ("attribute.name decs", "Invalid ordering key"),
-        ("name ACS", "Invalid ordering key"),
-        ("name decs", "Invalid ordering key"),
-        ("creation_timestamp DESC", "Invalid attribute key"),
-        ("last_updated_timestamp DESC blah", "Invalid order_by clause"),
-        ("", "Invalid order_by clause"),
-        ("timestamp somerandomstuff ASC", "Invalid order_by clause"),
-        ("timestamp somerandomstuff", "Invalid order_by clause"),
-        ("timestamp decs", "Invalid order_by clause"),
-        ("timestamp ACS", "Invalid order_by clause"),
-        ("name aCs", "Invalid ordering key"),
-    ],
-)
-def test_invalid_order_by_search_registered_models(order_by, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.parse_order_by_for_search_registered_models(order_by)
-    assert error_message in e.value.message
+    with pytest.raises(MlflowException, match=_get_regex_error_message(error_message)):
+        SearchRunsUtils.parse_order_by_for_search_runs(order_by)
 
 
 @pytest.mark.parametrize(
@@ -650,29 +480,10 @@ def test_invalid_order_by_search_registered_models(order_by, error_message):
     ],
 )
 def test_space_order_by_search_runs(order_by, ascending_expected):
-    (identifier_type, identifier_name, ascending,) = SearchUtils.parse_order_by_for_search_runs(
+    (identifier_type, identifier_name, ascending,) = SearchRunsUtils.parse_order_by_for_search_runs(
         order_by
     )
     assert identifier_type == "metric"
-    assert identifier_name == "Mean Square Error"
-    assert ascending == ascending_expected
-
-
-@pytest.mark.parametrize(
-    "order_by, ascending_expected",
-    [
-        ("tags.`Mean Square Error`", True),
-        ("tags.`Mean Square Error` ASC", True),
-        ("tags.`Mean Square Error` DESC", False),
-    ],
-)
-def test_space_order_by_search_registered_model(order_by, ascending_expected):
-    (
-        identifier_type,
-        identifier_name,
-        ascending,
-    ) = SearchUtils.parse_order_by_for_search_registered_models(order_by)
-    assert identifier_type == "tag"
     assert identifier_name == "Mean Square Error"
     assert ascending == ascending_expected
 
@@ -738,7 +549,7 @@ def test_pagination_for_runs(page_token, max_results, matching_runs, expected_ne
     encoded_page_token = None
     if page_token:
         encoded_page_token = base64.b64encode(json.dumps(page_token).encode("utf-8"))
-    paginated_runs, next_page_token = SearchUtils.paginate_runs(
+    paginated_runs, next_page_token = SearchRunsUtils.paginate_runs(
         runs, encoded_page_token, max_results
     )
 
@@ -767,6 +578,5 @@ def test_pagination_for_runs(page_token, max_results, matching_runs, expected_ne
     ],
 )
 def test_invalid_page_tokens_for_runs(page_token, error_message):
-    with pytest.raises(MlflowException) as e:
-        SearchUtils.paginate_runs([], page_token, 1)
-    assert error_message in e.value.message
+    with pytest.raises(MlflowException, match=_get_regex_error_message(error_message)):
+        SearchRunsUtils.paginate_runs([], page_token, 1)
